Bug 21587 - Check if common record has been deleted
Checking if the common record was deleted was only done in the rare case
where there also was a deleted enrichment or local record.
Now the check has been moved up a level so the local record action is 
only hit if the library either doesn't use enrichment or the common has
never existed.

git-svn-id: https://svn.dbc.dk/repos/updateservice/trunk@119310 b87a6194-a221-0410-ad6f-3b8ec35bba13