Fail on validation error
If the agency id in a record (001 *b) is not identical to the group id
in the request when sending a FBS/PH record to updateservice then that
request must fail with a validation error.

The validation error was in place but it didn't trigger a failed
validation. This meant the response was something like 'OK! But you are
not allowed to do that'. Now the request fails properly.