Remove links
If links are only removed if there is a parent record in 014 then
authority links won't ever be reset if a parent isn't detected.
Instead links should always be removed and then rebuild.

Right now there is problem with finding parent links with field 016 but
that will be solved later.