Clean dir on exit
Previously cleaning was done at the beginning of the job. A consequence
from this is that the checked out source code is still in the workspace
after the job is done. This might be the cause of runaway jobs when a
change to the meta-deploy project is submitted.
If the meta-deploy dir is no longer in the workspace this hopefully
won't happen.