Compare record with existing record when handling note and subject fields
In order to determine if a record can be updated by other libraries than
DBC we should look at the existing record and not the incoming record.
While the national common record logic was being looked at the check
for the codes was updated as well.

If a record can have its note and subject field updated then must
override the existing note/subject fields with the incoming record and
mark the fields as owned by the new library.