Prevent change from head to single record if the record has children
Updating head/section/volume/single volume is controlled by templates.

The type of record is controlled by the template. However there is
nothing that prevents someone from updating a head record with a single
record template which then allows the record to change type. So this we
have to check for in updateservice.