Do not add *r if it already exist
Previously the code always added a *r with the former ISBN. However this
meant a new *r was added everything the record was processed by update.
This is clearly an error, so it has been fixed.

Note that the test case number jumps from 2 to 4 because I have test
case 3 in a shelved changelist.