Add tests for checking if head volume is a text or sound record
The trigger for the pre-processing should be if the record has 520 *n.
If so check if the record has 009 *a a|r. If not check if head record
has 009 *r a|r. If the record is a text or sound record then look at the
record linked in 520 *n. The ISBN might be hiding in head volume there
as well.