Update modified date for all child records when authority record is updated
When a A-record is updated all direct child B-records should have their
modified date (001 *c) updated with current date. The easiest way to
update 001 *c is simply by storing the record and then StoreRecordAction
will take care of setting the modified date.

The children to the A-record children should be enqueued but not
updated.