Update child records only if there is a relevant change in the A-record
If all B-records have their modified date updated every time an A-record
changes it will cause way to many records to be printed during proof
printing. In order to avoid this B-records should only be updated if the
A-record contains a relevant change for proof printing.

Those fields are 100, 400 and 500.

Note: this solution cannot handle changing the order of a field. If that
is a necessary feature it will be implemented at a later time.