Return better error message when parent is missing
Previously the PreProcessingAction assumed that the parent record always
existed if such was defined. However this resulted in a confusing
JAXB.unmarshal exception when it wasn't the case because we then tried
to parse an empty byte[].

To fix this we now check to see if the parent record exist before
retrieving it, and if the parent doesn't exist a validation message is
returned.