Handle JSON values as String
We have started getting JSON replies from solr where the marc.001a value
can be a single string and not always a list. This causes a weird
exception which hits the user and prevents updating the record.
The problem is solved by checking if the JSON value is of type array or
string and handling the value accordingly.