Use external network instead
With internal network it became complicated to run multiple instances of
the compose file as the networks would either have to be autogenerated
(which might result in an invalid range) or conflict with other users
using the same range. Hence it seems like the original external network
actually was the best solution.