Move request and response object to business function
The trackingId is set when MDC content is created and that doesn't happen until the business function is called.
In order to get trackingId on the first and last log lines we either have to hack something together and set the MDC context before the business function are called or let the business function handle the raw request and response objects.

The latter seems like the easier choice.