Validate note and subject fields when not allowed to use those fields
Instead of validating that non-note and -subjects field are unchanged we now check if those specific fields are changed when the agency does not have permission to change them.
This means the record is validated ok if there are allowed changes to note and subject field in addition to illegal changes to other fields as the other changes are ignored when updating the record.

Auditors: mib, jbn