Perform extra check for ffu agency
The getFFULibraries call only return active FFU agencies. It must be possible to create a common record if there are deleted FFU agencies for the record id.
To fix this an extra call is made for the agency not found in the FFU list to check if the agency is actually FFU anyway.

Auditors: jbr, hwha, mib