Don't expect OK from warm up
We must provide username and password for the warm up request. But we can't provide real values because the code is public open source. And we can't implement some sort of hack (e.g. a specific username and password) that allows bypassing the IDP check, again because the code is public which means someone could potentially discover the hack and use it for malicious purpose.
So the easier way to do this, is to provide a bogus username and password and then expect an error message in the reply. The goal of initializing updateservice is still achieved.

In addition, the isReady check is made static.